Add support for script templates
For shell scripts it's not clear if it has a shebang line which differs
from the interpreter being used; e.g. we presume to use `sh` for all .sh
files, but they may have a #!/bin/bash shebang line. The only way to
honor the shebang is to write the contents out to a new file and execute
that file directly with `sh` rather than using `sh -c` on it.